Refactor to split out BoardRow
