+++ src/components/board/Board.tsx
@@ -1,14 +1,14 @@
 import React, { MouseEvent } from "react";
 import classNames from "classnames";
+import { createSlot } from "react-slotify";
+
 import { HighlightColors } from "../../types/player";
 import { GridIndex, GridData, GridItem } from "../../types/grid";
 import { chunkArray } from "../../utils";
 import { gridSize } from "../../const/gridData";
-import BoardCell from "../board-cell";
-import { createSlot } from "react-slotify";
+import { BoardRow, BoardRowCellSlot } from "./BoardRow";
 
 import styles from "./Board.module.scss";
-import { CellSlot } from "../board-cell/BoardCell";
 
 interface BoardProps {
   layout?: "grid" | "lined";
@@ -20,54 +20,9 @@ interface BoardProps {
   onClick: (e: MouseEvent, index: GridIndex) => void;
 }
 
-interface BoardRowProps {
-  gridSize: number;
-  gridItems: GridData;
-  indexRow: GridIndex;
-  highlighted?: GridIndex[];
-  highlightColor?: HighlightColors;
-  onClick: (e: MouseEvent, index: GridIndex) => void;
-}
-
-export const BoardRowCellSlot = createSlot<{
-  value: GridItem;
-  index: number | undefined;
-}>();
-
-const BoardRow = (props: React.PropsWithChildren<BoardRowProps>) => (
-  <tr className="BoardRow">
-    {props.gridItems.map((gridItem: GridItem, indexColumn: GridIndex) => {
-      const flatIndex = props.indexRow * props.gridSize + indexColumn;
-      return (
-        <td key={`board-row-cell-${flatIndex}`} className="BoardRow__cell">
-          <BoardCell
-            id={flatIndex}
-            value={gridItem}
-            isHighlighted={props.highlighted?.includes(flatIndex)}
-            highlightColor={props.highlightColor}
-            onClick={(e: MouseEvent) => props.onClick(e, flatIndex)}
-          >
-            {props.children && (
-              <CellSlot>
-                {(args) => (
-                  <BoardRowCellSlot.Renderer
-                    index={flatIndex}
-                    childs={props.children}
-                    value={args.value}
-                  />
-                )}
-              </CellSlot>
-            )}
-          </BoardCell>
-        </td>
-      );
-    })}
-  </tr>
-);
-
 export const BoardCellSlot = createSlot<{
   value: GridItem;
-  index: number | undefined;
+  index: GridIndex;
 }>();
 
 const Board = (props: React.PropsWithChildren<BoardProps>) => {
@@ -94,7 +49,7 @@ const Board = (props: React.PropsWithChildren<BoardProps>) => {
               indexRow={indexRow}
               highlighted={props.highlighted}
               highlightColor={props.highlightColor}
-              onClick={(e: MouseEvent, index: GridIndex) =>
+              onClick={(e: any, index: GridIndex) =>
                 !props.disabled && props.onClick(e, index)
               }
             >
